# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.PermissionStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - PermissionStatusActive
    - PermissionStatusInactive
...
  model.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  model.AssignIDsRequest:
    properties:
      menu_ids:
        example:
        - "1"
        - "2"
        items:
          type: string
        type: array
      permission_ids:
        example:
        - "1"
        - "2"
        items:
          type: string
        type: array
      role_ids:
        example:
        - "1"
        - "2"
        items:
          type: string
        type: array
    type: object
...
  service.CreatePermissionRequest:
    properties:
      action:
        maxLength: 50
        type: string
      code:
        maxLength: 100
        type: string
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        type: string
      resource:
        maxLength: 100
        type: string
      status:
        $ref: '#/definitions/entity.PermissionStatus'
    required:
    - action
    - code
    - name
    - resource
    type: object
...
  service.UpdatePermissionRequest:
    properties:
      action:
        maxLength: 50
        type: string
      code:
        maxLength: 100
        type: string
      description:
        maxLength: 500
        type: string
      id:
        type: string
      name:
        maxLength: 100
        type: string
      resource:
        maxLength: 100
        type: string
      status:
        $ref: '#/definitions/entity.PermissionStatus'
    required:
    - action
    - code
    - id
    - name
    - resource
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/coze-dev/alice
  description: Alice 企业级后端 API 文档。
  title: Alice API
  version: "1.0"
paths:
...
  /permissions:
    get:
      description: 分页获取权限列表
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 权限列表
      tags:
      - Permissions
    post:
      consumes:
      - application/json
      description: 创建一个新的权限
      parameters:
      - description: 创建权限请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 创建权限
      tags:
      - Permissions
  /permissions/{id}:
    delete:
      description: 删除指定权限
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除权限
      tags:
      - Permissions
    get:
      description: 根据ID获取权限详情
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取权限详情
      tags:
      - Permissions
    put:
      consumes:
      - application/json
      description: 更新指定权限
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新权限请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 更新权限
      tags:
      - Permissions
...
  /roles/{id}/permissions:
    delete:
      consumes:
      - application/json
      description: 从角色移除一组权限
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      - description: 权限ID集合 (permission_ids)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AssignIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 移除角色权限
      tags:
      - RolePermissions
    get:
      description: 获取指定角色的权限列表
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取角色权限
      tags:
      - RolePermissions
    post:
      consumes:
      - application/json
      description: 为指定角色批量分配权限
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      - description: 权限ID集合 (permission_ids)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AssignIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 为角色分配权限
      tags:
      - RolePermissions
...
  /users/{user_id}/permissions:
    get:
      description: 获取指定用户的权限列表
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取用户权限
      tags:
      - UserPermissions
  /users/{user_id}/permissions/check:
    get:
      description: 检查用户是否拥有某资源操作权限
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: string
      - description: 资源标识
        in: query
        name: resource
        required: true
        type: string
      - description: 操作标识
        in: query
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIResponse'
      security:
      - BearerAuth: []
      summary: 检查用户权限
      tags:
      - UserPermissions
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: '请输入形如: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"